+++ lib/util/extensions/datetime.dart
@@ -1,9 +1,14 @@
-import 'package:timeago/timeago.dart' as timeago;
+import 'package:flutter/material.dart';
+import 'package:timeago/timeago.dart';
 
 extension FancyTime on DateTime {
   /// returns `this` time as a relative, human-readable string. In short format
-  String get fancyShort => timeago.format(this, locale: 'en_short');
+  String timeagoShort(BuildContext context) => format(
+        this,
+        locale: '${Localizations.localeOf(context).toLanguageTag()}_short',
+      );
 
   /// returns `this` time as a relative, human-readable string
-  String get fancy => timeago.format(this);
+  String timeago(BuildContext context) =>
+      format(this, locale: Localizations.localeOf(context).toLanguageTag());
 }