+++ lib/widgets/post/post_info_section.dart
@@ -100,7 +100,10 @@ class PostInfoSection extends StatelessWidget {
                                   post.creator,
                                 ),
                         ),
-                        TextSpan(text: ' · ${post.post.published.fancyShort}'),
+                        TextSpan(
+                          text:
+                              ' · ${post.post.published.timeagoShort(context)}',
+                        ),
                         if (post.post.locked) const TextSpan(text: ' · 🔒'),
                         if (post.post.stickied) const TextSpan(text: ' · 📌'),
                         if (post.post.nsfw) const TextSpan(text: ' · '),