+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Changed
+
+- "Time ago" strings are now localized
+
 ## v0.7.0 - 2021-11-04
 
 ### Added